libcamera: simple: Increase the default number of streams to 2
In order to allow providing a raw stream together with a processed
stream, we need two streams.  If only one stream is eventually
requested, it doesn't harm to allocate two.

This is a hack for the lack of a better easy option.  The number of
streams is determined as a camera property in the pipeline matching.
The actual number of streams needed (one or two) is determined only when
examining roles in SimplePipelineHandler::generateConfiguration.

It's not obvious what to do better about the number of the streams.  It
seems that hardware pipelines utilize separate hardware streams while
with software ISP we have only a single input for multiple outputs.  A
fixed number of streams also doesn't scale but is good enough for our
current use case, which is supporting a single processed + a single raw
stream, not more.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>