libcamera: simple: Add plain output configurations to software ISP
When software ISP is active, only software ISP adjusted output
configurations are considered.  We want to support raw streams together
with software ISP and for that purpose, we need to track also unmodified
configurations.

A flag is added to SimpleCameraData::Configuration indicating whether
it's for software ISP.  Configurations for unmodified output sizes and
formats are added now whether software ISP is active or not.  We later
filter formats and output sizes for particular stream configuration
according to the new configuration flag.  This is just preparation and
raw output is still not supported; for this reason, we just look whether
software ISP is active; this will be changed in followup patches.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>