libcamera: simple: Protect against null maxPipeConfig
SimpleCameraData::pipeConfig_ is set to the determined maxPipeConfig if
no better configuration is found.  In the current code, maxPipeConfig
should be always set.  But it may be easy to miss that requirement and
end up with null maxPipeConfig on contingent code changes.

Let's add a check for nullptr to prevent segmentation fault surprises.
It doesn't harm.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>