libcamera: simple: Consider raw output configurations
If there is no software ISP, all the output formats and sizes are
relevant.  If there is software ISP, take only software ISP
configurations if there is no raw stream and only raw
configurations (otherwise it wouldn't be possible to produce a correct
passthrough output) if there is a raw stream.

TODO: This change looks dubious but it works.  Some form of filtering is
definitely needed otherwise the followup overlap check leads to mess and
incorrect output.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>