apps: cam: Write raw file if PPM cannot be written
`cam' application can be requested to write its output as a PPM file.
However, this is supported only for certain formats (only BGR888
currently).  If the output format cannot be written, `cam' reports an
error and doesn't write anything.

This is all right with a single stream input but impractical with
multiple streams of different output formats (e.g. a processed stream +
a raw stream) where some of them can be written in the supported format
while the other not.  In such a case, it's better to write the supported
formats as PPM files and the unsupported formats as raw files, with the
corresponding warning on stderr.  `.raw' extension is added to the file
name in such a case to make clear it's not a PPM file.

This is a sort of hack but serves the purpose for the moment.

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>