ipa: rkisp1: Add WDR algorithm
Add a WDR algorithm to do global tone mapping. Global tone mapping is
used to increase the perceived dynamic range of an image. The typical
effect is that in areas that are normally overexposed, additional
structure becomes visible.

The overall idea is that the algorithm applies an exposure value
correction to underexpose the image to the point where only a small
number of saturated pixels is left. This artificial underexposure is
then mitigated by applying a tone mapping curve.

This algorithm implements 4 tone mapping strategies:
- Linear
- Power
- Exponential
- Histogram equalization

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v3:
- Removed the need for a separate regulation loop, by not relying on an
  added ExposureValue but by applying a constraint to AGC regulation and
deducing the required WDR gain from the histogram. This makes the
structure easier and hopefully less prone to oscillations.
- Dropped debug metadata as this needs more discussions and is not
  required for the algorithm to work.
- Dropped minExposureValue as it is not used anymore.
- Added a damping on the gain applied by the WDR curve
- Moved strength into activeState so that it is reset on configure()

Changes in v2:
- Fixed default value for min bright pixels
- Added check for supported params type
- Reset PID controller
- Various fixes from Pauls review