libcamera: pipeline: uvcvideo: Fix `ExposureTimeMode` control
`ControlInfo(Span<const int32_t>{...})` calls the incorrect constructor
of `ControlInfo`. The intended constructor to be called is
`ControlInfo(Span<const ControlValue>, ...)` however that is not called
because a span of `const int32_t` is passed. Instead, the constructor
`ControlInfo(const ControlValue &min, const ControlValue &max, ...)`
will be called.

To fix this, simply pass a span of `ControlValue` instead.

Furthermore, the mapping in `UVCCameraData::processControl()` is also not
entirely correct because the control value is retrieved as a `bool`
instead of - the correct type - `int32_t`. Additionally, the available
modes are not taken into account.

To fix this, stores the available modes for `V4L2_CID_EXPOSURE_AUTO`
and select the appropriate mode based on the mapping established
in the comment.

Fixes: bad8d591f8acfa ("libcamera: uvcvideo: Register ExposureTimeMode control")
Signed-off-by: Barnabás Pőcze <pobrn@protonmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>