libcamera: thread: Fix event dispatcher creation
Use an appropriate compare-and-exchange atomic instruction
to set the event dispatcher pointer. This ensures that there
is only ever a single event dispatcher associated with a
given thread.

Consider e.g. the following series of events:

1. thread 0 :: if (!data_->dispatcher_.load(std::memory_order_relaxed))
               // reads nullptr and proceeds into the `if`'s block

2. thread 1 :: if (!data_->dispatcher_.load(std::memory_order_relaxed))
               // reads nullptr and proceeds into the `if`'s block

3. thread 0 :: data_->dispatcher_.store(new EventDispatcherPoll(), std::memory_order_release);
               // creates dispatcher and sets pointer

4. thread 1 :: data_->dispatcher_.store(new EventDispatcherPoll(), std::memory_order_release);
               // creates dispatcher and sets pointer

At the end, one of the event dispatchers is leaked, and
furthermore the `eventDispatcher()` calls might return
different values in the calling threads.

Signed-off-by: Barnabás Pőcze <pobrn@protonmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>