config: Look up log file in configuration file
Let's allow specifying the log file in the configuration file.  This
requires change in LIBCAMERA_LOG_FILE interpretation so that we can
disable the log file in isolated processes.  If LIBCAMERA_LOG_FILE is
defined but empty, it means there is no log file, even if defined in the
configuration file.  This should cause no compatibility trouble since
opening a file with an empty name makes little sense.

The configuration snippet:

  configuration:
    log:
      file: FILENAME

Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>