ipa: rkisp1: Allow processStats() to be called without stats buffer
When there are no stats available for a frame, it still makes sense to
call processStats() to fill in the metadata of that frame. This
mechanism is already used for the raw path. Allow it's use for non-raw
also.

The current code never produces buffers with id 0, but it is not
enforced. Add a assert to enforce that.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>

---

Changes in v2:
- Added an assert to ensure there is no buffer with id 0