ipa: rkisp1: Lazy initialise frame context
For per frame control we want to tick the IPA by the sensor frame
sequence instead of the request frame sequence. This has the side effect
that the IPA must be able to cope with situations where a frame context
is required for a frame that was not queued before (computeParams is
called without a corresponding request) or processStats is called for an
unexpected sequence number (because a scratch buffer was used on kernel
side)

Prepare for that by allowing the frame context to be initialized on demand.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>