libcamera: pipeline: uvcvideo: Fix `ExposureTimeMode` control setting
The mapping in `UVCCameraData::processControl()` is not entirely correct
because the control value is retrieved as a `bool` instead of `int32_t`.
Additionally, the available modes are not taken into account.

To fix this, retrieve the control with the proper type - `int32_t` -,
and use the available modes stored in `UVCCameraData` to determine
which value to use for the V4L2 control.

Fixes: bad8d591f8acfa ("libcamera: uvcvideo: Register ExposureTimeMode control")
Signed-off-by: Barnabás Pőcze <pobrn@protonmail.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>