libcamera: vector: Add a Span based constructor
When one wants to create a Vector from existing data, currently the only
way is via std::array. Add a Span based constructor to allow creation
from std::vectors and alike.

While at it, replace the manual loop with std::copy.

Signed-off-by: Stefan Klug <stefan.klug@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>

---

Changes in v2:
- Added this patch

Changes in v3:
- Improved commit message
- Added constexpr to the contructor
- Pass Span by value
- Removed initializer of data_ member which was only needed for the
  default constructor to be actually constexpr. But Vector is not used
as constexpr anywhere, so we can leave that as is (This will fix itself
with C++20).