libcamera: base: log: Take `LogCategory` by reference
When no log category is specified, `nullptr` is passed,
and then the `_log()` function implementations replace that
with `LogCategory::defaultCategory()`. But since the call site
always knows the log category, this condition can be removed
and the `_LOG1()` macro can use `LogCategory::defaultCategory()`.

So remove the condition from the `_log()` implementations and
use references to refer to log categories.

Signed-off-by: Barnabás Pőcze <barnabas.pocze@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>