libtuning: Gradient: Correct Linear distribution
The Linear distributor attempts to distribute pixels into sectors
of even size. Where the image width doesn't allow this it attempts
to create all but one sector of full size, and allows the caller
to choose to have the stunted sector at the front or the back. The
implicit assumption is that

	domain == (sector_size * (n_sectors - 1)) + remainder

which does not necessarily hold true. For example with 32 sectors and
a domain of 648 the calculated sector size will be 21 pixels, which
leads to 31 * 21 = 651 which is larger than the domain size.

Correct the issue by checking if there's more one stunted sector to
be filled. If there is, rather than following the remainder hint for
distribution place one of the stunted sectors at the front and one at
the back.

Signed-off-by: Daniel Scally <dan.scally@ideasonboard.com>
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>